fix: rename SendOnChainTransaction -> Send
