+++ src/screens/Wallets/Send/FeeRate.tsx
