+++ src/screens/Wallets/Send/Scanner.tsx
